#14 move: 컨벤션에 맞춰 폴더 구조 이동, 수정
