+++ src/app/components/addActivities/bannerImageInput/BannerImageInput.module.scss
