+++ src/app/components/addActivities/bannerImageInput/BannerImageInput.tsx
@@ -1,9 +1,10 @@
 import Image from 'next/image';
+import { useState } from 'react';
+import { FieldError, FieldErrorsImpl, Merge } from 'react-hook-form';
 import S from './BannerImageInput.module.scss';
 import plusIcon from '@/images/plus-icon.svg';
 import deleteButton from '@/images/delete-button.svg';
-import { FieldError, FieldErrorsImpl, Merge } from 'react-hook-form';
-import { useState } from 'react';
+
 interface BannerImageInputProps {
   error?: FieldError | Merge<FieldError, FieldErrorsImpl>;
   message?: string | FieldError | Merge<FieldError, FieldErrorsImpl>;
