+++ src/app/components/addActivities/subImageInput/SubImageInput.module.scss
