+++ src/app/components/addActivities/textarea/Textarea.module.scss
@@ -1,5 +1,4 @@
-@import '../../../styles/abstracts/core';
-
+@use '@/styles/abstracts/core' as *;
 .textareaWrapper {
   display: flex;
   flex-direction: column;
